+++ Assets/Scripts/Cards/CardTypes/Action/CardActionChipLotto.cs
@@ -96,6 +96,6 @@ public class CardActionChipLotto : CardMaster, ICardAction
 
     public override string GetDescription()
     {
-        return GameSettings.AddIcon(string.Format(GameSettings.LocalizeText(card_description), probability, (int)amount, damage, (int)mana));
+        return GameSettings.AddIcon(string.Format(GameSettings.LocalizeText(card_description), probability, (int)amount, (int)damage, (int)mana));
     }
 }
