+++ Assets/Scripts/Cards/CardTypes/Action/CardActionSkeletons.cs
@@ -47,8 +47,8 @@ public class CardActionSkeletons : CardMaster, ICardAction
     public void TriggerAction(CardMaster card, Transform target)
     {
         GameEvents.instance.TriggerActionCard(card, target);
-        // Use only parent variables: damage, health, mana, amount, etc.
-        // Example: Summon skeletons using amount as count, damage, health, mana
+        // Use only parent variables: damage, health, (int)mana, (int)amount, etc.
+        // Example: Summon skeletons using amount as count, (int)damage, health, (int)mana
         int count = Mathf.Max(1, (int)amount); // Use amount as skeleton count
         GameObject prefab = skeletonPrefab;
         if (prefab == null) return;
@@ -71,6 +71,6 @@ public class CardActionSkeletons : CardMaster, ICardAction
 
     public override string GetDescription()
     {
-        return GameSettings.AddIcon(string.Format(GameSettings.LocalizeText(card_description), damage, health, (int)amount));
+        return GameSettings.AddIcon(string.Format(GameSettings.LocalizeText(card_description), (int)damage, health, (int)amount));
     }
 }
