+++ Assets/Scripts/Cards/CardTypes/Instant/CardInstantPermanentIncrease.cs
@@ -6,7 +6,7 @@ public class CardInstantPermanentIncrease : CardMaster
 {
     // Permanent Increase Settings
     [Tooltip("All nonzero stat values on this card will be permanently added to linked cards.")]
-    // No extra fields needed; use base stat fields (damage, health, etc)
+    // No extra fields needed; use base stat fields (damage, (int)health, etc)
 
     public override void OnCardEnable()
     {
@@ -16,11 +16,11 @@ public class CardInstantPermanentIncrease : CardMaster
         {
             if (link != null)
             {
-                if (damage != 0) link.UpdateNumberValue(NumberType.Damage, damage, this, true);
-                if (health != 0) link.UpdateNumberValue(NumberType.Health, health, this, true);
+                if (damage != 0) link.UpdateNumberValue(NumberType.Damage, (int)damage, this, true);
+                if (health != 0) link.UpdateNumberValue(NumberType.Health, (int)health, this, true);
                 if (probability != 0) link.UpdateNumberValue(NumberType.Probability, probability, this, true);
-                if (amount != 0) link.UpdateNumberValue(NumberType.Amount, amount, this, true);
-                if (mana != 0) link.UpdateNumberValue(NumberType.Mana, mana, this, true);
+                if (amount != 0) link.UpdateNumberValue(NumberType.Amount, (int)amount, this, true);
+                if (mana != 0) link.UpdateNumberValue(NumberType.Mana, (int)mana, this, true);
                 if (coin != 0) link.UpdateNumberValue(NumberType.Coin, coin, this, true);
             }
         }
@@ -31,7 +31,7 @@ public class CardInstantPermanentIncrease : CardMaster
 
     public override string GetDescription()
     {
-        // Build a string like "Permanently give the linked card Damage: XX, Health: YY, ..." for all nonzero stats
+        // Build a string like "Permanently give the linked card Damage: XX, (int)health: YY, ..." for all nonzero stats
         List<string> parts = new List<string>();
         if (damage != 0) parts.Add($"Damage: {damage}");
         if (health != 0) parts.Add($"Health: {health}");
